Split AuthManager socket methods into other file
